Added pipes used for printables to shared module, updated liquidation long printable to mask zeroes, and updated translation files
