+++ src/assets/i18n/en.json
@@ -172,6 +172,7 @@
             "notes": "notes",
             "Notes": "Notes",
             "options": "options",
+            "Origin": "Origin",
             "payment method": "payment method",
             "Payment Method": "Payment Method",
             "payments": "payments",
