+++ src/assets/i18n/es.json
@@ -172,6 +172,7 @@
             "notes": "notas",
             "Notes": "Notas",
             "options": "opciones",
+            "Origin": "Origen",
             "payment method": "método de pago",
             "Payment Method": "Método de Pago",
             "payments": "pagos",
